+++ core/dbt/contracts/graph/compiled.py
@@ -1,18 +1,18 @@
 from dbt.contracts.graph.parsed import (
+    HasTestMetadata,
     ParsedNode,
     ParsedAnalysisNode,
-    ParsedModelNode,
+    ParsedDataTestNode,
     ParsedHookNode,
+    ParsedModelNode,
     ParsedResource,
     ParsedRPCNode,
+    ParsedSchemaTestNode,
     ParsedSeedNode,
     ParsedSnapshotNode,
     ParsedSourceDefinition,
-    ParsedTestNode,
     SeedConfig,
     TestConfig,
-    TestMetadata,
-    PARSED_TYPES,
 )
 from dbt.node_types import NodeType
 from dbt.contracts.util import Replaceable
@@ -42,7 +42,6 @@ class CompiledNode(ParsedNode):
     extra_ctes_injected: bool = False
     extra_ctes: List[InjectedCTE] = field(default_factory=list)
     injected_sql: Optional[str] = None
-    wrapped_sql: Optional[str] = None
 
     def prepend_ctes(self, prepended_ctes: List[InjectedCTE]):
         self.extra_ctes_injected = True
@@ -116,11 +115,19 @@ class CompiledSnapshotNode(CompiledNode):
 
 
 @dataclass
-class CompiledTestNode(CompiledNode):
+class CompiledDataTestNode(CompiledNode):
+    resource_type: NodeType = field(metadata={'restrict': [NodeType.Test]})
+    config: TestConfig = field(default_factory=TestConfig)
+
+
+@dataclass
+class CompiledSchemaTestNode(CompiledNode, HasTestMetadata):
     resource_type: NodeType = field(metadata={'restrict': [NodeType.Test]})
     column_name: Optional[str] = None
     config: TestConfig = field(default_factory=TestConfig)
-    test_metadata: Optional[TestMetadata] = None
+
+
+CompiledTestNode = Union[CompiledDataTestNode, CompiledSchemaTestNode]
 
 
 def _inject_ctes_into_sql(sql: str, ctes: List[InjectedCTE]) -> str:
@@ -183,26 +190,43 @@ def _inject_ctes_into_sql(sql: str, ctes: List[InjectedCTE]) -> str:
     return str(parsed)
 
 
-COMPILED_TYPES: Dict[NodeType, Type[CompiledNode]] = {
-    NodeType.Analysis: CompiledAnalysisNode,
-    NodeType.Model: CompiledModelNode,
-    NodeType.Operation: CompiledHookNode,
-    NodeType.RPCCall: CompiledRPCNode,
-    NodeType.Seed: CompiledSeedNode,
-    NodeType.Snapshot: CompiledSnapshotNode,
-    NodeType.Test: CompiledTestNode,
+PARSED_TYPES: Dict[Type[CompiledNode], Type[ParsedResource]] = {
+    CompiledAnalysisNode: ParsedAnalysisNode,
+    CompiledModelNode: ParsedModelNode,
+    CompiledHookNode: ParsedHookNode,
+    CompiledRPCNode: ParsedRPCNode,
+    CompiledSeedNode: ParsedSeedNode,
+    CompiledSnapshotNode: ParsedSnapshotNode,
+    CompiledDataTestNode: ParsedDataTestNode,
+    CompiledSchemaTestNode: ParsedSchemaTestNode,
 }
 
 
-def compiled_type_for(parsed: ParsedNode):
-    if parsed.resource_type in COMPILED_TYPES:
-        return COMPILED_TYPES[parsed.resource_type]
+COMPILED_TYPES: Dict[Type[ParsedResource], Type[CompiledNode]] = {
+    ParsedAnalysisNode: CompiledAnalysisNode,
+    ParsedModelNode: CompiledModelNode,
+    ParsedHookNode: CompiledHookNode,
+    ParsedRPCNode: CompiledRPCNode,
+    ParsedSeedNode: CompiledSeedNode,
+    ParsedSnapshotNode: CompiledSnapshotNode,
+    ParsedDataTestNode: CompiledDataTestNode,
+    ParsedSchemaTestNode: CompiledSchemaTestNode,
+}
+
+
+# for some types, the compiled type is the parsed type, so make this easy
+CompiledType = Union[Type[CompiledNode], Type[ParsedResource]]
+
+
+def compiled_type_for(parsed: ParsedNode) -> CompiledType:
+    if type(parsed) in COMPILED_TYPES:
+        return COMPILED_TYPES[type(parsed)]
     else:
         return type(parsed)
 
 
 def parsed_instance_for(compiled: CompiledNode) -> ParsedResource:
-    cls = PARSED_TYPES.get(compiled.resource_type)
+    cls = PARSED_TYPES.get(type(compiled))
     if cls is None:
         # how???
         raise ValueError('invalid resource_type: {}'
@@ -212,22 +236,33 @@ def parsed_instance_for(compiled: CompiledNode) -> ParsedResource:
     return cls.from_dict(compiled.to_dict(), validate=False)
 
 
-# This is anything that can be in manifest.nodes and isn't a Source.
-NonSourceNode = Union[
+NonSourceCompiledNode = Union[
     CompiledAnalysisNode,
+    CompiledDataTestNode,
     CompiledModelNode,
     CompiledHookNode,
     CompiledRPCNode,
+    CompiledSchemaTestNode,
     CompiledSeedNode,
     CompiledSnapshotNode,
-    CompiledTestNode,
+]
+
+NonSourceParsedNode = Union[
     ParsedAnalysisNode,
+    ParsedDataTestNode,
     ParsedModelNode,
     ParsedHookNode,
     ParsedRPCNode,
+    ParsedSchemaTestNode,
     ParsedSeedNode,
     ParsedSnapshotNode,
-    ParsedTestNode,
+]
+
+
+# This is anything that can be in manifest.nodes and isn't a Source.
+NonSourceNode = Union[
+    NonSourceCompiledNode,
+    NonSourceParsedNode,
 ]
 
 # We allow either parsed or compiled nodes, or parsed sources, as some