+++ core/dbt/parser/macros.py
@@ -63,7 +63,7 @@ class MacroParser(BaseParser[ParsedMacro]):
             try:
                 ast = jinja.parse(block.full_block)
             except CompilationException as e:
-                e.node = base_node
+                e.add_node(base_node)
                 raise e
 
             macro_nodes = list(ast.find_all(jinja2.nodes.Macro))